+++ services/catalog/src/simcore_service_catalog/api/dependencies/user_groups.py
@@ -4,7 +4,7 @@ from fastapi import Depends, Query
 from models_library.groups import GroupAtDB
 from models_library.users import UserID
 
-from ...db.repositories.groups import GroupsRepository
+from ...repositories.groups import GroupsRepository
 from .database import get_repository
