+++ services/catalog/src/simcore_service_catalog/core/background_tasks.py
@@ -25,10 +25,10 @@ from simcore_service_catalog.services import manifest
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncEngine
 
-from ..db.repositories.groups import GroupsRepository
-from ..db.repositories.projects import ProjectsRepository
-from ..db.repositories.services import ServicesRepository
 from ..models.services_db import ServiceAccessRightsAtDB, ServiceMetaDataDBCreate
+from ..repositories.groups import GroupsRepository
+from ..repositories.projects import ProjectsRepository
+from ..repositories.services import ServicesRepository
 from ..services import access_rights
 
 _logger = logging.getLogger(__name__)
@@ -115,9 +115,9 @@ async def _ensure_registry_and_database_are_synced(app: FastAPI) -> None:
     director_api = get_director_api(app)
     services_in_manifest_map = await manifest.get_services_map(director_api)
 
-    services_in_db: set[
-        tuple[ServiceKey, ServiceVersion]
-    ] = await _list_services_in_database(app.state.engine)
+    services_in_db: set[tuple[ServiceKey, ServiceVersion]] = (
+        await _list_services_in_database(app.state.engine)
+    )
 
     # check that the db has all the services at least once
     missing_services_in_db = set(services_in_manifest_map.keys()) - services_in_db
