+++ services/catalog/src/simcore_service_catalog/repositories/groups.py
@@ -6,9 +6,9 @@ from models_library.groups import GroupAtDB
 from pydantic import TypeAdapter
 from pydantic.types import PositiveInt
 
-from ...exceptions.errors import UninitializedGroupError
-from ..tables import GroupType, groups, user_to_groups, users
+from ..exceptions.errors import UninitializedGroupError
 from ._base import BaseRepository
+from ._tables import GroupType, groups, user_to_groups, users
 
 
 class GroupsRepository(BaseRepository):
