+++ services/catalog/src/simcore_service_catalog/services/access_rights.py
@@ -1,6 +1,4 @@
-""" Services Access Rights policies
-
-"""
+"""Services Access Rights policies"""
 
 import logging
 import operator
@@ -17,9 +15,9 @@ from pydantic.types import PositiveInt
 from sqlalchemy.ext.asyncio import AsyncEngine
 
 from ..api.dependencies.director import get_director_api
-from ..db.repositories.groups import GroupsRepository
-from ..db.repositories.services import ServicesRepository
 from ..models.services_db import ServiceAccessRightsAtDB
+from ..repositories.groups import GroupsRepository
+from ..repositories.services import ServicesRepository
 from ..utils.versioning import as_version, is_patch_release
 
 _logger = logging.getLogger(__name__)
