+++ services/catalog/tests/unit/test_services_compatibility.py
@@ -11,8 +11,8 @@ from models_library.users import UserID
 from packaging.specifiers import SpecifierSet
 from packaging.version import Version
 from pytest_mock import MockerFixture, MockType
-from simcore_service_catalog.db.repositories.services import ServicesRepository
 from simcore_service_catalog.models.services_db import ReleaseDBGet
+from simcore_service_catalog.repositories.services import ServicesRepository
 from simcore_service_catalog.services.compatibility import (
     _get_latest_compatible_version,
     evaluate_service_compatibility_map,
